requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
python-chess==1.999